requirements.txt
fastapi~=0.115
uvicorn[standard]~=0.30
pydantic~=2.9
python-dotenv~=1.0

langchain~=0.3
langchain-openai~=0.2
openai~=1.55

SQLAlchemy~=2.0

psycopg[binary]~=3.2
langsmith~=0.4